third_party/blink/web_tests/external/wpt/svg/linking/scripted/a.text-getter-01-expected.txt

This is a testharness.js-based test.
PASS SVGAElement.text getting
FAIL Test for anchor 0 assert_equals: expected (string) "a b c " but got (undefined) undefined
FAIL Test for anchor 1 assert_equals: expected (string) "a b c " but got (undefined) undefined
FAIL Test for anchor 2 assert_equals: expected (string) "a b c " but got (undefined) undefined
FAIL Test for anchor 3 assert_equals: expected (string) "a b c " but got (undefined) undefined
FAIL Test for anchor 4 assert_equals: expected (string) "a b c " but got (undefined) undefined
Harness: the test ran to completion.
